+++ Stackage/UpdateBuildPlan.hs
@@ -32,6 +32,7 @@ updateBuildConstraints BuildPlan {..} =
     bcBuildToolOverrides = bpBuildToolOverrides
     bcTellMeWhenItsReleased = mempty -- we don't care when doing an update
     bcNoRevisions = bpNoRevisions
+    bcCabalFormatVersion = bpCabalFormatVersion
 
     bcPackageConstraints name = PackageConstraints
         { pcVersionRange = addBumpRange (maybe anyVersion pcVersionRange moldPC)